added timeouts for ear movements
One of the ears on my Karotz robot stopped moving.  The callbacks will
not fire until the ear completes a full revolution.  This was causing
the app to get stuck waiting for the ear to complete a full rotation.

Added a timeout at various points so that the ear rotation will not
block.